update
Изменено создание основных окон, добавлены графики статистики, улучшена статистика набора, добавлен тест на скорости клика, добавлен локальный чат

+++ FH/chart.h
@@ -8,7 +8,7 @@ public:
         int day;
         int speed;
         int amount;
-        int mistakes;
+        float mistakes;
 };
